+++ src/wine-host/bridges/vst3.cpp
@@ -1211,13 +1211,11 @@ size_t Vst3Bridge::register_object_instance(
                     },
                     [&](YaAudioProcessor::Process& request)
                         -> YaAudioProcessor::Process::Response {
-                        // HACK: When a plugin doesn't handle denormals
-                        //       properly, we can force the FTZ flag to be set
-                        //       to work around this
-                        std::optional<ScopedFlushToZero> ftz_guard =
-                            config.force_ftz
-                                ? std::make_optional<ScopedFlushToZero>()
-                                : std::nullopt;
+                        // Most plugins will already enable FTZ, but there are a
+                        // handful of plugins that don't that suffer from
+                        // extreme DSP load increases when they start producing
+                        // denormals
+                        ScopedFlushToZero ftz_guard;
 
                         // As suggested by Jack Winter, we'll synchronize this
                         // thread's audio processing priority with that of the